REFACTORING: - Added Labolatory level to Research time to build dependence - Added NaniteFactory level to Ship time to build dependence

+++ QuantumWorld_v1.0/ViewModel/ResearchViewModel.cs
@@ -14,7 +14,7 @@ namespace QuantumWorld_v1._0.ViewModel
     {
         private PlayerModel _player;
 
-        int timeToEnd;
+        float timeToEnd;
         bool isBusy;
 
         DispatcherTimer researchTimer;
